Datasets/create_RAG_data/rag_requirements.txt
chromadb
tiktoken
sentence-transformers
langchain
langchain-openai
python-dotenv
pydantic
